Fix things getting wrapped in Stores instead of ThingStores (which has a del method). Move store to their own file.
